% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affineCipher.R
\name{affineCipher}
\alias{affineCipher}
\title{Encrypt a string using a shift cipher}
\usage{
affineCipher(plainText, stretch, shift)
}
\arguments{
\item{plainText}{A string of lowercase letters}

\item{stretch}{An integer used to stretch the function.}

\item{shift}{An integer (mod 26) used to shift.}
}
\value{
An encrypted (or decrypted) string, where each letter has been shifted by \code{shift} and stretched by \code{stretch} places.
}
\description{
Encrypt a string using a shift cipher
}
\examples{
plainText <- "dog"
cipherText <- affineCipher(plainText,3,7)
print(cipherText)
You cannot decrypt a cipherText using the encryptAffineCipher function
}

