% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vigenere Cipher.R
\name{countMatches}
\alias{countMatches}
\title{Count matches within a text.}
\usage{
countMatches(ct, shift)
}
\arguments{
\item{ct}{a string of crypted text}

\item{shift}{an integer to shift by}
}
\value{
a table that counts the amount of matches
}
\description{
Count matches within a text.
}
\examples{
countMatches("hello", 5)

}

