% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decryptAES.R
\name{decryptAES}
\alias{decryptAES}
\title{AES decrypter}
\usage{
decryptAES(msg, key)
}
\arguments{
\item{msg}{The encrypted message.}

\item{key}{The key in string form (NOT RAW)}
}
\value{
A decrypted message.
}
\description{
AES decrypter
}
\examples{



}

