% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseAffineCipher.R
\name{decryptAffineCipher}
\alias{decryptAffineCipher}
\title{Decrypt a string with a known affine cipher parameters}
\usage{
decryptAffineCipher(cryptedText, stretch, shift)
}
\arguments{
\item{stretch}{An integer (mod 26) to stretch by}

\item{shift}{An integer (mod 26) to shift by}

\item{cryptedtext}{A string of lowercase letters}
}
\value{
A decrypted string, where each letter has been shifted by \code{shift} and stretched \code{stretch}
}
\description{
Decrypt a string with a known affine cipher parameters
}
\examples{
plainText <- "cat"
cipherText <- "smr"
print(plaintext)
}

