% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterFrequencies.R
\name{digramTable}
\alias{digramTable}
\title{Creates a Digram table from a text.}
\usage{
digramTable(txt)
}
\arguments{
\item{txt}{The text you wish the function to analyze.}
}
\value{
a table with the number of letter combinations
}
\description{
Creates a Digram table from a text.
}
\examples{
digramTable("hellohowareyou")
}

