% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FermatTest.R
\name{FermatTest}
\alias{FermatTest}
\title{Determines if a number is "probably" Prime}
\usage{
FermatTest(n, a)
}
\arguments{
\item{n}{A bigz integer.}

\item{a}{Some bigz integer less than n.}
}
\value{
A boolean value.
}
\description{
Checks if a bigz integer, n, is "probably" prime using Fermat's theorem.
}
\examples{
FermatTest(as.bigz(9^9), as.bigz(9^8))


}

