% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vigenere Cipher.R
\name{findVigKey}
\alias{findVigKey}
\title{Find Vigenere Key}
\usage{
findVigKey(ciphertext, keyLength)
}
\arguments{
\item{ciphertext}{a vigenere encrypted string}

\item{keyLength}{the length of the key to the vigenere cipher}
}
\value{
an integer that is the length of the string.
}
\description{
Find Vigenere Key
}
\examples{
findVigKey("zyupzvjslelzkovevveobtwkjoqdlvokqyskvoqemkmdczjyacsdyrdcyretadwcrefxlyrefxlmyklbsdxzfqftqjwvdeskzoncsmancy", 6)
}

