% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hillCipher.R
\name{hillCipher}
\alias{hillCipher}
\title{Encrypts messages according to the Hill Cipher algorithm.}
\usage{
hillCipher(txt, keyMatrix)
}
\arguments{
\item{txt}{The plaintext represented as a string you would like to enrypt.}

\item{keyMatrix}{A matrix that will be used to encrypt the plaintext.}
}
\value{
A string
}
\description{
Encrypts messages according to the Hill Cipher algorithm.
}
\examples{
hillCipher("hello", matrix(c(1,2,3,4,5,6,7,8,9), nrow=3, ncol=3, byrow=TRUE))

}

