% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterFrequencies.R
\name{letterCounts}
\alias{letterCounts}
\title{Count the Letters}
\usage{
letterCounts(txt)
}
\arguments{
\item{txt}{A string you would like to analyze}
}
\value{
A table that describes the amount of repeated letters you have in your txt
}
\description{
Count the Letters
}
\examples{
letterCounts("hello")
}

