% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vigenere Cipher.R
\name{letterFreq}
\alias{letterFreq}
\title{Find the letter frequency of a text.}
\usage{
letterFreq(txt)
}
\arguments{
\item{txt}{a string to analyze}
}
\value{
a vector of frequencies for each letter of the alphabet (in alphabetical order)
}
\description{
Find the letter frequency of a text.
}
\examples{
letterFreq("hello")
}

