% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_logs.R
\name{nextPrimRoot}
\alias{nextPrimRoot}
\title{Next Primitive Root}
\usage{
nextPrimRoot(p, after = 1)
}
\arguments{
\item{p}{A prime number. Could be \code{bigz}.}

\item{after}{Number after which to start testing. Defaults to 1.}
}
\value{
The smallest primitive root modulo \code{p} greater than \code{after}. Applies
the procedure described in Exercise 21 on p. 107 of [Trappe].
}
\description{
Finds the next primitive root after \code{after} modulo \code{p}.
}
\examples{
nextPrimRoot(601) # should be 7
nextPrimRoot(nextprime(as.bigz("1203481092840918409408098")), 200) # should be 203
}

