% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneTimePad.R
\name{oneTimePad}
\alias{oneTimePad}
\title{One Time Pad function will encrypt and decrypt messages according to 1 time pad algorithm}
\usage{
oneTimePad(pt, padkey)
}
\arguments{
\item{pt}{The plaintext, represented as a string, that you would like to encrypt.}

\item{padkey}{A key that is a string and is equal in length to the plaintext.}
}
\value{
A string.
}
\description{
One Time Pad function will encrypt and decrypt messages according to 1 time pad algorithm
}
\examples{
oneTimePad("james", "kdlaf")

}

