% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vigenere Cipher.R
\name{shiftVec}
\alias{shiftVec}
\title{Shift Vector}
\usage{
shiftVec(v, n)
}
\arguments{
\item{v}{Takes a vector v}

\item{n}{an integer n to shift each element of the vector}
}
\value{
a vector
}
\description{
Shift Vector
}
\examples{
shiftVec(c(1,1,2,3), 5)
}

