% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vigenere Cipher.R
\name{skipString}
\alias{skipString}
\title{Skip String}
\usage{
skipString(txt, n, r)
}
\arguments{
\item{txt}{a string you wish to analyze}

\item{n}{an integer that represents your initial position}

\item{r}{an integer you wish to skip through the string by}
}
\value{
a string of letters that were r length apart
}
\description{
Skip String
}
\examples{
skipString("hellohowareyou", 1, 2)
}

