% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterFrequencies.R
\name{vigenere}
\alias{vigenere}
\title{Encrypt using a Vigenere Cipher.}
\usage{
vigenere(txt, keyVector)
}
\arguments{
\item{txt}{A string you would like to encrypt.}

\item{keyVector}{A vector to encrypt your string.}
}
\value{
An encrypted string
}
\description{
Encrypt using a Vigenere Cipher.
}
\examples{
vigenere("helloworld", "at")
}

